pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nest"
version = "0.1.0"
description = "Nested Embedding Search Tool — single-file semantic search. SQLite for embeddings."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Brenner Cruvinel", email = "brenner@notlikeus.dev" },
]
keywords = ["embeddings", "semantic-search", "faiss", "sqlite", "archive"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
]
dependencies = [
    "numpy>=1.24",
    "faiss-cpu>=1.7",
    "zstandard>=0.20",
]

[project.optional-dependencies]
encode = ["sentence-transformers>=2.2"]
dev = ["pytest>=7.0", "sentence-transformers>=2.2"]

[project.scripts]
nest = "nest.cli:main"

[project.urls]
Repository = "https://github.com/brennercruvinel/nest"

[tool.hatch.build.targets.wheel]
packages = ["src/nest"]
